rename to lazysearch, add styling changes and add fontawesome to index.html
