+++ index.html
@@ -5,6 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/src/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite App</title>
+    <script src="https://kit.fontawesome.com/4ff3c29809.js" crossorigin="anonymous"></script>
   </head>
   <body>
     <div id="root"></div>
