+++ src/lazysearch/index.tsx
